# vector_bench: remove randomness